refactor(consultoria): split landing sections
